pyproject.toml
# ABOUTME: Project configuration for Arklex Investment Research Agent demo
# ABOUTME: Extends FinRobot with UBS House View RAG and multi-turn Q&A capabilities

[project]
name = "investor-agent"
version = "0.1.0"
description = "Investment Research Q&A Agent for Arklex demo"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Core FinRobot dependencies
    "pyautogen>=0.2.19",
    "ipython>=8.12.0",

    # LLM and embedding
    "openai>=1.0.0",
    "langchain>=0.1.20",
    "langchain-openai>=0.0.5",
    "langchain-community>=0.0.20",

    # Vector database
    "chromadb>=0.4.0",

    # Financial data sources
    "yfinance>=0.2.0",
    "finnhub-python>=2.4.0",
    "sec-api>=1.0.0",
    "fredapi>=0.5.0",

    # PDF processing
    "pymupdf>=1.23.0",
    "pypdf2>=3.0.0",
    "pdfplumber>=0.10.0",

    # Data handling
    "pandas>=2.0.3",
    "numpy>=1.26.4",

    # Web UI
    "gradio>=4.0.0",

    # API
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",

    # Utilities
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "tenacity>=8.3.0",
    "tqdm>=4.66.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"
